requirements.txt
asgiref==3.3.4
astroid==2.5.3
colorama==0.4.4
Django==3.2
django-advanced-filters==1.3.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-braces==1.14.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-widget-tweaks==1.4.8
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pylint==2.7.4
pytz==2021.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
wrapt==1.12.1
